WIP: rendering react components from markdown, rendering images from src
